+++ src/ts/types/beatmap/wrapper/lightColorEventBoxGroup.ts
@@ -7,5 +7,5 @@ export interface IWrapLightColorEventBoxGroup<
     TBase extends Record<keyof TBase, unknown> = Record<string, unknown>,
     TFilter extends Record<keyof TFilter, unknown> = Record<string, unknown>,
 > extends IWrapEventBoxGroup<TGroup, TBox, TBase, TFilter> {
-    events: IWrapLightColorEventBox<TBox, TBase, TFilter>[];
+    boxes: IWrapLightColorEventBox<TBox, TBase, TFilter>[];
 }